pyproject.toml
[tool.poetry]
name = "django_excercises"
version = "0.1.0"
description = ""
authors = ["jyothi <vishnu.vutukuri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
psycopg2-binary = "^2.9.1"
django-countries = "^7.2.1"
django-phonenumber-field = "^5.2.0"
phonenumbers = "^8.12.32"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
